requirements.txt
mysql-connector-python==9.1.0
bcrypt==4.2.0
